Clear up a large number of NSInteger/NSUInteger issues
- a few places used NSInteger in method declaration, but int in the implemntation:
  on 64bit these are different sized types
- using %d, %u for NSInteger/NSUInteger

+++ src/Amazon.Runtime/AmazonAbstractJsonWebServiceClient.m
@@ -53,7 +53,7 @@
                     
         [self startSyncRequest:generatedRequest forRequest:urlRequest response:response originalRequest:originalRequest];
         
-        AMZLogDebug(@"Response Status Code : %d", response.httpStatusCode);
+        AMZLogDebug(@"Response Status Code : %ld", (long)response.httpStatusCode);
         if ( [self shouldRetry:response exception:((AmazonRequestDelegate *)generatedRequest.delegate).exception]) {
             AMZLog(@"Retring Request: %d", retries);
