+++ src/Amazon.Runtime/AmazonAbstractWebServiceClient.m
@@ -112,7 +112,7 @@
         }
         
         [self startSyncRequest:generatedRequest forRequest:urlRequest response:response originalRequest:originalRequest];
-        AMZLogDebug(@"Response Status Code : %d", response.httpStatusCode);
+        AMZLogDebug(@"Response Status Code : %d", (int)response.httpStatusCode);
         if ( [self shouldRetry:response] ) {
             AMZLog(@"Retring Request: %d", retries);
             
@@ -208,7 +208,7 @@
     return NO;
 }
 
--(void)pauseExponentially:(NSInteger)tryCount
+-(void)pauseExponentially:(int)tryCount
 {
     NSTimeInterval pause = self.delay * (pow(2, tryCount));
