+++ src/Amazon.S3/Model/S3UploadInputStream.m
@@ -62,7 +62,7 @@
 -(NSInteger)read:(uint8_t *)buffer maxLength:(NSUInteger)len
 {
 	NSInteger nBytesRead = [stream read:buffer maxLength:(kS3UploadInputStreamChunkSize * packetSize)];
-	AMZLogDebug(@"S3UploadInputStream: read %d bytes (%d max)", nBytesRead, len);
+	AMZLogDebug(@"S3UploadInputStream: read %ld bytes (%lu max)", (long)nBytesRead, (unsigned long)len);
 	[NSThread sleepForTimeInterval:delay];
 	return nBytesRead;
 }
